requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
pydantic-settings==2.8.1
python-telegram-bot==21.10
requests==2.32.3
python-multipart==0.0.20
jinja2==3.1.5
